change to parameterized tests

+++ src/main/java/ru/netology/sqr/SQRService.java
@@ -1,7 +1,7 @@
 package ru.netology.sqr;
 
 public class SQRService {
-    public long calcSqrt(int x, int y) {
+    public int calcSqrt(int x, int y) {
         int q = 0;
         // где х - нижняя граница диапазона, у - верхняя, q - количество чисел
         for (int i = 10; i <= 99; i++) {
